+++ ha-addon/calendhd/rootfs/opt/calendhd/public/_app/immutable/entry/start.DAgiD_36.js
@@ -1,1 +1,0 @@
-import{r as e,s as t}from"../chunks/C45GsEeF.js";export{t as load_css,e as start};